backend/requirements/base.txt
Django>=6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
asgiref>=3.7.0
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
Pillow>=10.0.0
aiofiles>=23.0.0
httpx>=0.25.0